Improve templates; Require login on tracker creation

+++ db/migrate/20200122224141_create_trackers.rb
@@ -3,7 +3,7 @@ class CreateTrackers < ActiveRecord::Migration[6.0]
     create_table :trackers do |t|
       t.string :title, null: false
       t.string :url, null: false
-      t.float :threshold_price, null: false
+      t.integer :threshold_price, null: false
       t.boolean :enabled, default: true
       t.references :user, null: false
